+++ Utility.h
@@ -3,4 +3,4 @@
 static const int ScreenWidth = 1920;
 static const int ScreenHeight = 1080;
 
-static const int Lanes[5] = { 272, 616, 960, 1304, 1648 };
+static const float Lanes[5] = { 272, 616, 960, 1304, 1648 };
